requirements.txt
manage.py==0.2.10
flask-bcrypt==0.7.1
flask-login==0.5.0
flask-migrate==2.5.3
flask-sqlalchemy==2.4.4
flask-static-digest==0.1.3
flask-wtf==0.14.3
flask==1.1.2
sqlalchemy==1.3.20
python-dotenv==0.14.0
mysqlclient==2.0.1
